Move getResourceLinks and getMeta to runtime/web
