+++ src/runtime/web/util.js
@@ -1,3 +1,18 @@
+// A document.ready shim
+// https://github.com/whatwg/html/issues/127
+export function documentReady() {
+  if (document.readyState !== 'loading') {
+    return Promise.resolve();
+  }
+
+  return new Promise(resolve => {
+    document.addEventListener('readystatechange', function onrsc() {
+      document.removeEventListener('readystatechange', onrsc);
+      resolve();
+    });
+  });
+}
+
 export function getResourceLinks(head) {
   return Array.prototype.map.call(
     head.querySelectorAll('link[rel="localization"]'),
